+++ src/views/location-detail/components/LocationHeader.vue
@@ -23,7 +23,14 @@ const { t } = useI18n()
         {{ location.description }}
       </p>
       <p v-if="location.updatedAt" class="mt-2 text-xs text-slate-500">
-        {{ $t('locations.lastUpdated', { date: $d(new Date(location.updatedAt), 'short') }) }}
+        {{
+          location.updatedByUsername
+            ? $t('locationDetail.header.lastUpdatedBy', {
+                date: $d(new Date(location.updatedAt), 'shortDateTime'),
+                username: location.updatedByUsername,
+              })
+            : $t('locations.lastUpdated', { date: $d(new Date(location.updatedAt), 'shortDateTime') })
+        }}
       </p>
       <button
         v-if="canManage"